新しいSavedTabs機能を追加し、カスタムモードとドメインモードのコンテナを実装しました。また、保存されたタブの状態管理を行うためのフックを追加し、ストレージの変更を同期するサービスを実装しました。
